+++ union_rag/data_annotation.py
@@ -2,13 +2,10 @@ import functools
 import itertools
 import json
 import os
-import random
 from dataclasses import dataclass
-from datetime import timedelta
-from typing import List, Annotated, Dict, Optional
+from typing import List, Annotated
 
-from flytekit import task, workflow, Resources, Artifact, Secret, map_task, current_context, wait_for_input
-from flytekit.core.artifact import Inputs
+from flytekit import task, workflow, Resources, Artifact, Secret, map_task, current_context
 from flytekit.types.file import FlyteFile
 
 from union_rag.document import CustomDocument
@@ -25,14 +22,6 @@ class QuestionAndAnswers:
     answers: List[str]
     url: str
 
-@dataclass
-class AnnotatedQuestions:
-    id: int
-    question: str
-    answer_1: str
-    answer_2: str
-    preferred_answer: Optional[str]
-    user_answer: Optional[str]
 
 
 @task(
@@ -59,7 +48,7 @@ def generate_qa_pairs(flyte_doc: CustomDocument, n_questions: int, n_answers: in
         template="""Given the following context, generate {n_questions} relevant and specific questions that require thoughtful, nuanced answers.
 
         Context: {context}
-        
+
         Requirements:
         1. Each question should be clear but encourage deep thinking or inference.
         2. Avoid questions that can be answered with a simple factual statement.
@@ -67,7 +56,7 @@ def generate_qa_pairs(flyte_doc: CustomDocument, n_questions: int, n_answers: in
         4. Ensure questions cover different aspects and perspectives of the context.
         5. The question should relate to the overarching theme or concepts in the context but not be directly answerable by it. Think of the question as a follow-up from an attentive student seeking to explore the topic further or clarify a complex point.
         6. IMPORTANT: Place each question on a new line, with no numbering or prefixes.
-        
+
         Questions:"""
     )
 
@@ -143,73 +132,9 @@ def create_dataset(questions_and_answers: List[List[QuestionAndAnswers]]) -> Fly
 
 
 @workflow
-def data_synthesis_workflow(documents: List[CustomDocument] = KnowledgeBase.query(), n_questions: int = 1, n_answers: int = 5) -> Annotated[FlyteFile, QuestionAndAnswerDataset]:
+def data_synthesis_workflow(documents: List[CustomDocument] = KnowledgeBase.query(), n_questions: int = 1,
+                            n_answers: int = 5) -> Annotated[FlyteFile, QuestionAndAnswerDataset]:
     partial_task = functools.partial(generate_qa_pairs, n_questions=n_questions, n_answers=n_answers)
     questions_and_answers = map_task(partial_task)(flyte_doc=documents)
     dataset = create_dataset(questions_and_answers=questions_and_answers)
     return dataset
-
-@task(
-    container_image=image,
-    requests=Resources(cpu="2", mem="4Gi"),
-    cache=True,
-    cache_version="1",
-)
-def sample_triple(random_seed: int, n_samples: int, dataset: FlyteFile) -> List[Dict]:
-    dataset.download()
-    with open(dataset.path, 'r') as f:
-        qa_triples_list = json.load(f)
-
-    assert len(
-        qa_triples_list) >= n_samples, f"Not enough samples in the dataset. Required: {n_samples}, Available: {len(qa_triples_list)}"
-
-    random.seed(random_seed)
-    sampled_qa = random.sample(qa_triples_list, k=n_samples)
-
-    return sampled_qa
-
-@task(
-    container_image=image,
-    requests=Resources(cpu="2", mem="4Gi"),
-)
-def append_annotated_data(feedback: str, master_annotation_dataset: FlyteFile) -> FlyteFile:
-    parsed_annotation_json = json.loads(feedback)
-
-    master_annotation_dataset.download()
-    with open(master_annotation_dataset.path, 'r') as file:
-        master_annotation_dataset_list = json.load(file)
-
-    master_annotation_dataset_list.extend(parsed_annotation_json)
-
-    print(master_annotation_dataset_list)
-
-    file_path = 'master_annotation_dataset.json'
-    with open(file_path, 'w') as f:
-        json.dump(master_annotation_dataset_list, f, indent=4)
-
-    return FlyteFile(path=file_path)
-
-@workflow
-def create_annotation_set(random_seed: int, n_samples: int = 10, dataset: FlyteFile = QuestionAndAnswerDataset.query(), annotation_set_name: str = DEFAULT_ANNOTATION_SET_NAME, master_annotation_dataset: FlyteFile = AnnotatedDataset.query(name=Inputs.annotation_set_name)) -> Annotated[FlyteFile, AnnotatedDataset(name=Inputs.annotation_set_name)]:
-    triple_sample = sample_triple(random_seed=random_seed, n_samples=n_samples, dataset=dataset)
-    feedback = wait_for_input("feedback", timeout=timedelta(hours=1), expected_type=str)
-    triple_sample >> feedback
-    annotated_dataset = append_annotated_data(feedback=feedback, master_annotation_dataset=master_annotation_dataset)
-    return annotated_dataset
-
-
-@task(container_image=image)
-def init_annotation_set() -> FlyteFile:
-    file_path = 'master_annotation_dataset.json'
-    with open(file_path, 'w') as f:
-        json.dump([], f, indent=4)
-
-    return FlyteFile(path=file_path)
-
-@workflow
-def init_annotation_set_wf(annotation_set_name: str = DEFAULT_ANNOTATION_SET_NAME) -> Annotated[FlyteFile, AnnotatedDataset(name=Inputs.annotation_set_name)]:
-    return init_annotation_set()
-
-
-if __name__ == "__main__":
-    data_synthesis_workflow()
